Extract shared SAS timestamp rules
